Add authorization to portfolio link

+++ app/Http/Controllers/Portfolio/PortfolioLinkController.php
@@ -3,15 +3,16 @@
 namespace App\Http\Controllers\Portfolio;
 
 use App\PortfolioUnit;
-use Illuminate\Http\Request;
+use App\Http\Requests\LinkRequest;
 use App\Link;
-use App\Http\Controllers\LinkController;
+use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Redirect;
 
-class PortfolioLinkController extends LinkController
+class PortfolioLinkController extends Controller
 {
     public function index(PortfolioUnit $portfolioUnit)
     {
+        $this->authorize('view', $portfolioUnit);
         $links = $portfolioUnit->other_links;
         $editRoute = route('portfolios.links.edit', ['portfolio_unit' => $portfolioUnit->pid, 'link' => 'LLIINNKK']);
         return view('portfolios.links.index', compact('portfolioUnit', 'links', 'editRoute'));
@@ -19,40 +20,42 @@ class PortfolioLinkController extends LinkController
 
     public function create(PortfolioUnit $portfolioUnit)
     {
+        $this->authorize('create', $portfolioUnit);
         $link = new Link;
         $formAction = route('portfolios.links.store', ['portfolio_unit' => $portfolioUnit->pid]);
-        return view('links.edit', compact('link', 'formAction'));
+        $parentModel = $portfolioUnit;
+        $deleteRoute = '';
+        return view('links.edit', compact('link', 'formAction', 'parentModel', 'deleteRoute'));
     }
 
-    public function store(Request $request, PortfolioUnit $portfolioUnit)
+    public function store(LinkRequest $request, PortfolioUnit $portfolioUnit)
     {
-        $validated = $this->validateValues($request);
-        $link = new Link($validated);
+        $this->authorize('view', $portfolioUnit);
+        $link = new Link($request->validated());
         $link->linkable_subtype = Link::SUBTYPE_PORTFOLIO_OTHER;
         $portfolioUnit->links()->save($link);
-
         return Redirect::route('portfolios.links.index', ['portfolio_unit' => $portfolioUnit->pid]);
     }
 
     public function edit(PortfolioUnit $portfolioUnit, Link $link)
     {
-        $this->validateModelLink($portfolioUnit, $link, Link::SUBTYPE_PORTFOLIO_OTHER);
+        $this->authorize('view', $portfolioUnit);
         $deleteRoute = route('portfolios.links.destroy', ['portfolio_unit' => $portfolioUnit->pid, 'link' => $link->pid]);
         $formAction = route('portfolios.links.update', ['portfolio_unit' => $portfolioUnit->pid, 'link' => $link->pid]);
-        return view('links.edit', compact('link', 'deleteRoute', 'formAction'));
+        $parentModel = $portfolioUnit;
+        return view('links.edit', compact('link', 'deleteRoute', 'formAction', 'parentModel'));
     }
 
-    public function update(Request $request, PortfolioUnit $portfolioUnit, Link $link)
+    public function update(LinkRequest $request, PortfolioUnit $portfolioUnit, Link $link)
     {
-        $this->validateModelLink($portfolioUnit, $link, Link::SUBTYPE_PORTFOLIO_OTHER);
-        $validated = $this->validateValues($request);
-        $link->update($validated);
+        $this->authorize('update', $portfolioUnit);
+        $link->update($request->validated());
         return Redirect::route('portfolios.links.index', ['portfolio_unit' => $portfolioUnit->pid]);
     }
 
     public function destroy(PortfolioUnit $portfolioUnit, Link $link)
     {
-        $this->validateModelLink($portfolioUnit, $link, Link::SUBTYPE_PORTFOLIO_OTHER);
+        $this->authorize('delete', $portfolioUnit);
         $link->delete();
         return Redirect::route('portfolios.links.index', ['portfolio_unit' => $portfolioUnit->pid]);
     }
